finished form refactoring
